Fix dead exhaustion sentinel in Cluster_Splitter::DeltaM
DeltaM declared its retry counter as `size_t trials = 1000` and exits the
draw loop on `(deltaM>deltaMmax) && (trials--)>0`. On genuine exhaustion the
final guard evaluates `(0)>0` (false, loop exits) but post-decrements `trials`
to SIZE_MAX, so the closing `return trials>0?deltaM:0.` takes the `deltaM`
branch and returns the last over-maximum draw instead of the intended `0.`.

Commit a461d7063 fixed the loop's trial count (`>1000` -> `>0`, so the cap is
now a real 1000 attempts rather than one), but the unsigned underflow left the
`:0.` exhaustion sentinel unreachable. Returning the over-max deltaM inflates
m_R2[i] in MakeLongitudinalMomentaMassSimple, which then fails
`m_R2[0]+m_R2[1]<m_Q2` and forces a retry instead of accepting a minimal-mass
split -- self-correcting but wasteful.

Declare `trials` as `long int`, matching the sibling
MakeLongitudinalMomentaMassSimple loop, so the counter exhausts to -1 and the
`0.` branch is reached.

+++ AHADIC++/Decays/Cluster_Splitter.C
@@ -213,7 +213,7 @@ double Cluster_Splitter::DeltaM(const size_t & cl) {
   double deltaM, deltaMmax = m_Q-sqrt(m_m2min[0])-sqrt(m_m2min[1]);
   double mean =  m_mean[cl], sigma = 1./(m_c[cl] * sqrt(m_kt02));
   double arg  =  1.-exp(-sigma * deltaMmax);
-  size_t trials = 1000;
+  long int trials = 1000;
   do {
     // Weibull distribution
     //deltaM = sqrt(offset+pow(-log(ran->Get()),1./m_a[cl])*lambda);
